RFC 8422: Basic ed25519/ed448 support in BcTlsCrypto

+++ tls/src/main/java/org/bouncycastle/tls/crypto/TlsCrypto.java
@@ -82,7 +82,7 @@ public interface TlsCrypto
      * @param signatureAlgorithm the algorithm of interest.
      * @return true if signatureAlgorithm is supported, false otherwise.
      */
-    boolean hasSignatureAlgorithm(int signatureAlgorithm);
+    boolean hasSignatureAlgorithm(short signatureAlgorithm);
 
     /**
      * Return true if this TlsCrypto can support the passed in signature algorithm.
